Create a separate operator for Single.cast and test it.
