+++ src/test/java/rx/internal/operators/OperatorCastTest.java
@@ -69,31 +69,4 @@ public class OperatorCastTest {
 
         ts.assertError(ClassCastException.class);
     }
-
-    @Test
-    public void testSingleCast() {
-        Single<?> source = Single.just(1);
-        Single<Integer> single = source.cast(Integer.class);
-
-        @SuppressWarnings("unchecked")
-        Observer<Integer> observer = mock(Observer.class);
-        single.subscribe(observer);
-        verify(observer, times(1)).onNext(1);
-        verify(observer, times(1)).onNext(1);
-        verify(observer, never()).onError(
-            org.mockito.Matchers.any(Throwable.class));
-        verify(observer, times(1)).onCompleted();
-    }
-
-    @Test
-    public void testSingleCastWithWrongType() {
-        Single<?> source = Single.just(1);
-        Single<Boolean> single = source.cast(Boolean.class);
-
-        @SuppressWarnings("unchecked")
-        Observer<Boolean> observer = mock(Observer.class);
-        single.subscribe(observer);
-        verify(observer, times(1)).onError(
-            org.mockito.Matchers.any(ClassCastException.class));
-    }
 }
